Fix:修改测试的bug

+++ app/src/main/java/com/moko/bxp/a/c/entity/AdvInfo.java
@@ -18,7 +18,7 @@ public class AdvInfo implements Serializable {
     public String deviceInfoMhz;
     public String deviceInfoTxPower;
     public int deviceInfoAdvInterval;
-    public double temperature;
+    public int temperature;
     public int alarmCount;
     public String alarmStatus;
     public int advType;
